# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cosco API Documentation",
        "description": "API documentation for Cosco application",
        "contact": {
            "email": "admin@cosco.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user's orders",
                "description": "Returns a list of all orders for the authenticated user",
                "operationId": "getUserOrders",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by order status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "paid",
                                "shipped",
                                "delivered",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "description": "Filter by payment status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "unpaid",
                                "paid",
                                "refunded"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Create a new order",
                "description": "Creates a new order with items for the authenticated user",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "سفارش با موفقیت ایجاد شد"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Address does not belong to user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "description": "Returns details of a specific order (only user's own orders)",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order-items/{id}": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Get a specific order item",
                "description": "Returns details of a specific order item (only user's own order items)",
                "operationId": "getOrderItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderItemResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not authorized to view this order item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateOrderRequest": {
                "title": "Create Order Request",
                "description": "Request schema for creating a new order",
                "required": [
                    "items",
                    "shipping_address_id"
                ],
                "properties": {
                    "items": {
                        "description": "Array of order items",
                        "type": "array",
                        "items": {
                            "required": [
                                "product_slug",
                                "quantity"
                            ],
                            "properties": {
                                "product_slug": {
                                    "description": "Product slug",
                                    "type": "string",
                                    "example": "mhsol-aol"
                                },
                                "quantity": {
                                    "description": "Quantity of the product",
                                    "type": "integer",
                                    "maximum": 100000,
                                    "minimum": 1,
                                    "example": 2
                                }
                            },
                            "type": "object"
                        }
                    },
                    "shipping_address_id": {
                        "description": "Shipping address ID",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "description": "Order description",
                        "type": "string",
                        "example": "توضیحات سفارش",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AddressResource": {
                "title": "Address Resource",
                "description": "User address resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "postal_code": {
                        "description": "Postal code",
                        "type": "string",
                        "example": "1234567890"
                    },
                    "plaque": {
                        "description": "Building plaque number",
                        "type": "string",
                        "example": "12"
                    },
                    "address": {
                        "description": "Full address",
                        "type": "string",
                        "example": "خیابان انقلاب، پلاک 12"
                    },
                    "phone": {
                        "description": "Phone number",
                        "type": "string",
                        "example": "02112345678"
                    },
                    "is_default": {
                        "description": "Is default address",
                        "type": "boolean",
                        "example": true
                    },
                    "province_title_fa": {
                        "description": "Province name in Persian",
                        "type": "string",
                        "example": "تهران"
                    },
                    "province_id": {
                        "description": "Province ID",
                        "type": "integer",
                        "example": 10
                    },
                    "city_title_fa": {
                        "description": "City name in Persian",
                        "type": "string",
                        "example": "تهران"
                    },
                    "city_id": {
                        "description": "City ID",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "OrderItemResource": {
                "title": "Order Item Resource",
                "description": "Order item resource schema",
                "properties": {
                    "quantity": {
                        "description": "Quantity of the product in order",
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "description": "Product price in Toman",
                        "type": "number",
                        "format": "decimal",
                        "example": 7500
                    },
                    "total_price": {
                        "description": "Total price (quantity × price) in Toman",
                        "type": "number",
                        "format": "decimal",
                        "example": 15000
                    },
                    "product_name": {
                        "description": "Product name",
                        "type": "string",
                        "example": "محصول نمونه"
                    },
                    "product_image": {
                        "description": "Product image URL",
                        "type": "string",
                        "example": "https://example.com/storage/products/image.jpg"
                    },
                    "product_slug": {
                        "description": "Product slug for URL",
                        "type": "string",
                        "example": "mhsol-aol"
                    }
                },
                "type": "object"
            },
            "OrderResource": {
                "title": "Order Resource",
                "description": "Order resource schema",
                "properties": {
                    "status": {
                        "description": "Order status label in Persian",
                        "type": "string",
                        "example": "در انتظار"
                    },
                    "total_amount": {
                        "description": "Total amount in Toman",
                        "type": "number",
                        "format": "decimal",
                        "example": 15000
                    },
                    "formatted_total_amount": {
                        "description": "Formatted total amount with currency",
                        "type": "string",
                        "example": "15,000 تومان"
                    },
                    "payment_status": {
                        "description": "Payment status label in Persian",
                        "type": "string",
                        "example": "پرداخت نشده"
                    },
                    "description": {
                        "description": "Order description",
                        "type": "string",
                        "example": "توضیحات مربوط به سفارش",
                        "nullable": true
                    },
                    "shipping_address": {
                        "$ref": "#/components/schemas/AddressResource"
                    },
                    "order_items": {
                        "description": "Array of order items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResource"
                        }
                    },
                    "received_at": {
                        "description": "Order received date in Persian calendar",
                        "type": "string",
                        "example": "1402-10-25",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Order creation date in Persian calendar",
                        "type": "string",
                        "example": "1402-10-25"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response schema",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object",
                        "example": []
                    },
                    "status": {
                        "type": "integer",
                        "example": 400
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}